Shift the git controller
Earlier, I was using BlueJ 4 which allowed to use
git from itself but I changed to BlueJ 3 which doesn't
support git and therefore, I have to use git bash.
This commit and probably the next will commit some
not so imp changes. I will add .gitignore in a future
commit.